+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.0.0-alpha.2
+
+- Added persistent SQLite project registry.
+- Added project, book, release, export and history models.
+- Added registry CLI and integrity tests.
+
 # Changelog
 
 ## [3.0.0-alpha.1] - 2026-08-03